+++ src/Components/Dashboard/Dashboard.tsx
@@ -7,7 +7,7 @@ import Manage from "../Manage/Manage";
 import Servers from "../Servers/Servers";
 import RiderStudio from "../Rider/RiderStudio";
 import Browse from "../Browse/Browse";
-import Shop from "../Shop/Shop";
+import ShopCatalog from "../Shop/ShopCatalog";
 import ModDetail from "../ModDetail/ModDetail";
 import Settings, { type SectionId } from "../Settings/Settings";
 import Tour, { TourContext, TOUR_DONE_KEY } from "../Tour/Tour";
@@ -129,7 +129,7 @@ const Dashboard = ({ welcomeActive = false }: DashboardProps) => {
               onChangeType={changeType}
             />
           ) : view === "shop" ? (
-            <Shop refreshKey={libraryVersion} />
+            <ShopCatalog />
           ) : view === "library" ? (
             <Library
               modType={modType}